Moving images to public folder
