+++ my-react-portfolio/public/linkedinIcon.png
