+++ my-react-portfolio/public/myImage.jpg
