+++ my-react-portfolio/src/projects.json
@@ -2,7 +2,7 @@
   "projects": [
     {
       "name": "Weather app",
-      "image": ["src/assets/projectPhotos/weatherApp.jpg"],
+      "image": ["../public/weatherApp.jpg"],
       "description": "Stay informed about the weather with this sleek and straightforward weather dashboard. Powered by data from the Open Weather Map API, this intuitive web application provides users with real-time weather updates along with a convenient four-day forecast. Easily access current weather conditions and plan ahead with confidence using this user-friendly weather dashboard.",
       "tags": ["HTML5", "CSS3", "React", "APIs", "Node"],
       "netlify": "https://weatherforecastbypaula.netlify.app/",
@@ -10,7 +10,7 @@
     },
     {
       "name": "Music Releases",
-      "image": ["src/assets/projectPhotos/musicReleases.jpg"],
+      "image": ["../public/musicReleases.jpg"],
       "description": "Explore the latest music releases with ease on this intuitive web page powered by React and Spotify's API. Discover new albums and single releases directly from Spotify, presented in a user-friendly interface designed for music enthusiasts. Stay updated on the latest music trends effortlessly with this interactive platform. ",
       "tags": ["HTML5", "CSS3", "React", "Node", "vite"],
       "netlify": "https://heroic-crisp-8a05fa.netlify.app/",
@@ -18,7 +18,7 @@
     },
     {
       "name": "Happy Thoughts",
-      "image": ["src/assets/projectPhotos/happyThoughts.jpg"],
+      "image": ["../public/happyThoughts.jpg"],
       "description": "Step into the realm of React as you embark on building a Twitter-inspired web platform. This project offers a hands-on opportunity to delve into React's state management intricacies while mimicking the functionality and feel of Twitter. Users can engage in real-time interactions, post updates and explore trending topics. ",
       "tags": ["HTML5", "CSS3", "React", "APIs", "Node", "React Hooks", "Vite"],
       "netlify": "https://kaleidoscopic-gumption-4a304b.netlify.app/",
@@ -26,7 +26,7 @@
     },
     {
       "name": "React Survey",
-      "image": ["src/assets/projectPhotos/survey.jpg"],
+      "image": ["../public/survey.jpg"],
       "description": "Embark on a React journey with our customizable survey form project, where React state and form control take center stage. Upon submission, users are treated to a comprehensive summary. Our design ethos emphasizes accessibility, prioritizing user-friendly experiences for all.",
       "tags": ["HTML5", "CSS3", "React", "Node", "Vite"],
       "netlify": "https://projectsurveyformreact.netlify.app/",
